Adding Rook
Now, Rook can move freely logic yet to be added. Will also be adding
some pics from a actual gameplay b/w me and computer in the comments
section of this commit (the game was played days before i started
coding during the research phase to refresh my memories and knowledge
of the game of chess which i used to play when i was about 15 years old.